Refactoring the Tree-based classifiers. Improve ShapletForest

+++ core/src/main/java/org/briljantframework/matrix/storage/DoubleStorage.java
@@ -43,7 +43,7 @@ public class DoubleStorage extends AbstractStorage {
 
   @Override
   public double getDouble(int index) {
-    return values[((int) index)];
+    return values[index];
   }
 
   @Override
